Extracts out article list logic and separates code in multiple files
